requirements.txt
Django==3.1.4
django-chartjs==2.2.1
django-crispy-forms==1.10.0
django-import-export==2.5.0